chore: move utils related to campus and cluster class under lib
