+++ web/ui/src/components/ClusterMap/index.ts
@@ -7,8 +7,4 @@ export {
   ClusterWorkspaceWithUser,
 } from './ClusterTableMap';
 export type { ClusterContainerProps, MapLocation } from './types';
-export {
-  extractNode,
-  extractandRemoveNode,
-  findCampusPerSafeLink,
-} from './utils';
+export { extractNode, extractandRemoveNode } from './utils';
