+++ web/ui/src/components/ClusterMap/utils.ts
@@ -1,6 +1,4 @@
-import { ICampus } from '@lib/clustersMap/types';
 import { NodeFinderFunc, NodeIndexFinderFunc } from './types';
-import Campuses from '@lib/clustersMap';
 
 /**
  * findIndexForNode will extract the index of the node object from a graphql
@@ -61,22 +59,4 @@ export const extractandRemoveNode: NodeFinderFunc = (
   connection.edges = connection.edges.slice(index, 1).concat([]);
 
   return location.node;
-};
-
-/**
- * findCampusPerSafeLink will find a campus object from its link representation.
- * For example, if you want to find the campus object for the campus "Paris",
- * you can use the link "paris" to find it.
- * @param slug - the link of the campus in the safe url format (e.g. "paris")
- * @returns the campus object or undefined if not found
- */
-export const findCampusPerSafeLink = (
-  slug: string | undefined,
-): ICampus | undefined => {
-  if (!slug) return undefined;
-
-  return Object.values(Campuses).find(
-    (iCampusObject) =>
-      iCampusObject.identifier().toSafeLink() === slug.toSafeLink(),
-  );
 };
