+++ web/ui/src/lib/clustersMap/index.ts
@@ -1,4 +1,5 @@
 export { countryEmoji } from './countryEmoji';
 export { Campuses, Campuses as default } from './campuses.generated';
+export { findCampusPerSafeLink } from './utils';
 export type { CampusIdentifier } from './types.generated';
 export type { ClusterMapEntity, ICampus, ICluster } from './types';
